requirements.txt
contextlib2==0.6.0.post1
cssselect==1.1.0
lxml==4.5.2
MouseInfo==0.1.3
mss==6.0.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PySide2==5.15.1
PyTweening==1.0.3
PyYAML==5.3.1
schema==0.7.2
shiboken2==5.15.1
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-win_amd64.whl
